Fix grapher 404 page
This is an alternative fix for the one reverted in 24f6288.
